rename now_playing.rs to player_context.rs
